Flask app with Python/requirements.txt
Click==8.0.4
Flask==2.0.3
Jinja2==3.0.3
MarkupSafe==1.1.1
python-dotenv==0.19.2
twilio==7.7.0
Werkzeug==2.0.3
requests==2.27.1
idna==3.3
urllib3==1.26.8
MarkupSafe==2.1.0